loader/requirements.txt
pyzmq
locustio